requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-grappelli==4.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
